+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Fix #2788: Support FIPS mode in kubernetes-client with BouncyCastleFipsProvider
 * Fix #2910: Move crd-generator tests from kubernetes-itests to kubernetes-tests
 * Fix #3005: Make it possible to select which CRD version is generated / improve output
+* Fix #3015: Thread interruption in a nominal case (like closing the client) are now logged in debug
 
 #### Dependency Upgrade
 * Fix #2979: Update Kubernetes Model to v1.21.0